colorbox: More work to speed up full frame grabs
In the generated c++ API (client and server) there were some
inefficiencies with the toJsonValue() and fromJsonValue() helper
functions when QByteArray objects were involved.

In the case of toJsonValue() it was needlessly creating a temp QString
and in the case of fromJsonValue() it was converting through a
std::string, which is slower than just keeping with QString.