colorbox: Bringing back the Preview Tap
The Preview Tap is back! It is now limited to only the Input and Output.

To change the Preview Tap to Input:
- The user can click the "INPUT" label in the Pipeline View
- The user can use the keyboard shortcut `ctrl + (`
- An OEM can use the API

To change the Preview Tap to Output:
- The user can click the "OUTPUT" label in the Pipeline View
- The user can use the keyboard shortcut `ctrl + )`
- An OEM can use the API

The 'Capture' button in the Frame Store setup area is now either
'Capture Input' or 'Capture Output' based on what tap the user has
selected.

The Shortcuts tab in System Settings has been updated with these 2 new
shortcuts.

This is for issue bridge/soji#247