colorbox: Adding a new generated client API
Adding the cpp-restsdk[0] client api, a cpp API that does not depend on
Qt was requested by a potential user.

Have a working get_temp sample application so far. Still need to go back
and put in the custom templates (like the other clients have) to support
min/max constraints.

[0]: https://openapi-generator.tech/docs/generators/cpp-restsdk/