colorbox: Remove CSC Filter control from System Settings->Controls in UI
The control is not needed at this time, we are going to default to not
using the filter for now (hard-coded in featureD).

The setting is remaining in the rest API for future use, added two new
enums to CSCFilter: Auto and None. None will be used as the default
when/if this is used in the future.

This addresses the remaining part of issue bridge/soji#245