colorbox: Added ability to select a preset to load at startup
Now the user can choose to load a certain preset at startup if they
want, the previous behavior of not loading a preset (that uses the last
state) is also available.

This feature is being added so presets that contain dynamic lut
information can be automatically loaded at startup so the system is put
back into the correct dynamic lut mode as left by a 3rd party
application.

Note that the preset will not be loaded until 5 seconds after the server
starts, this is done so featureD has time to startup so the websocket
connection can be established between server and featureD so the dynamic
lut information can be transferred. This wait time may need to be made
longer once we start playing with the new feature.

This addresses issue bridge/soji#256