colorbox: Fix the rgb_triplet cpprestsdk demo on Windows
Missed wrapping "rgbtriplet" in a to_string_t(), this is not a problem
on Mac and Linux since utility::string_t is a narrow string, whereas on
Windows it is a wide string.

Also had to tweak the parsing of the returned buffer so it works on
Windows, likely was due to the narrow/wide difference between the
platforms.