Updated the documentation
There is now a root level README.md that gives an overview of what this
repo is for, parts of this came out of the previous pdf document.

The bulk of the old pdf documentation is now in the cpp-qt/README.md, the
old pdf document has been removed.

Also fixed some grammar/wording and removed references to Soji and
"Color Box", replaced with "ColorBox".